refactor(cli): move reusable CLI internals into core modules (#341)
- Extract pure search policy / freshness / mode helpers into
  src/search-core.ts (SearchMode, AutoSearchMode, Staleness, computeStaleness,
  formatFreshnessFooter, auto-threshold helpers, explain-empty diagnostics).
- Extract error classification / format helpers into src/search-errors-core.ts
  (full move; src/cli-search-errors.ts is deleted).
- cli-search.ts, cli-promote.ts, cli-remember.ts shed test-only and
  internal-only exports (cli-search.ts 19 -> 11, cli-promote.ts 24 -> 13,
  cli-remember.ts 12 -> 2).
- retrieval-eval.ts and ingest-quarantine.ts now import from search-core /
  search-errors-core instead of CLI adapters, fixing the boundary inversion
  noted in #341. A new src/search-core.boundaries.test.ts pins the invariant.

Closes #341
