+++ src/cli-stats.ts
@@ -6,7 +6,7 @@ import {
   exitCodeForFailure,
   formatKbSearchFailureJson,
   formatKbSearchFailureStderr,
-} from './cli-search-errors.js';
+} from './search-errors-core.js';
 import { loadManagerForModel, loadWithJsonRetry } from './cli-shared.js';
 import { FaissIndexManager } from './FaissIndexManager.js';
 import { computeKbStats, type ComputeKbStatsOptions, type KbStatsPayload } from './kb-stats.js';
